pyproject.toml
[project]
name = "secureorder-pro-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.5",
    "httpx>=0.28.1",
    "langchain-core>=1.2.11",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.8",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
